Unlock or error on locked model
When the tool is run from the GUI and the model is locked, the user will be asked whether or not they want to unlock the library and proceed with obfuscation. If the tool is run from the command line on a locked model, an error will be reported to the command window.